[flow] More straightforward class self type
Summary:
Currently, we make the self type of a class to be a potentially polymorphic inner class type. However, this is transformed into an instance type as the bound of this typeparam. This is overly complicated, since what we want is fixpoint type that satisfies:

```
inner class instance type = InstanceT which contains this_tparam
this_tparam: GenericT(bound=inner class instance type)
```

We achieve this currently by (using non-polymorphic case for simplicity):

```
self=ClassT(ThisInstanceT(inner class instance type))
this_tparam: GenericT(bound=Flow.mk_instance self)
```

This diff depends on the `ThisInstanceT` built in previous diff, so that we can avoid the `Flow.mk_instance` that's hard to do in a world where the self type is a lazy OpenT that can only be forced after the entire class type is constructed.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52970904

fbshipit-source-id: 84504c4715e25bc7d83ae433b7e166d5c2c6bad9