[flow] The fixpoint miracle
Summary:
This is the diff where the magic happens. After all the preparation in previous diffs, the construction of class types no longer requires the use of any unresolved tvars, so we can tie the knot of the class self type using a fixpoint trick, similar to the one done in D44725347.

After this, the class types are fully resolved, which opens the door for us to move some of the class/instance type related constraints out of the giant pattern match in flow_js.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52970905

fbshipit-source-id: 0c66dc8b78d9c2795b9ff685ef57724b0fada54e