[flow][refactor] Move ESM exports' mutation on module info out of statement.ml
Summary:
In this diff, I refactor all the calls to import_export modules into a new module module_exports_analyzer. Later, this module will be used as a single place where the module type will be generated. Even later in the future, it will also be used to generate a type for a namespace.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53031652

fbshipit-source-id: f18f7fb1b7904940b43e794478c2d5319066a3ec