[flow] Remove useless write and read of declare module names
Summary:
This pair of read and write is needed back when we still use type_inference_js code path to generate libdef types, so we need to simulate a read for all toplevel names to get the type.

However, now that libdefs types are generated with type sig, this is no longer necessary, so let's just remove it. I found it when I'm looking into namespace support.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53068449

fbshipit-source-id: 1e5b850b102043a8de910e426a5db5eb4a0081c6