[flow][as-const] boilerplate support in type-sig
Summary:
In type-sig the as-const information wraps a `Type_sig.value`, and is a `value` itself. It is generated in Type_sig_parse as we parse the respective AST node, and it is applied on the payload `value` in Type_sig_merge. The latter part has not been implemented yet. It will be added in the following diffs that add support for each kind of literal.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D53076954

fbshipit-source-id: 43adbaf2bda7033ebb1c97488b5e2a29b5f5397c