[flow] Add test of fundamental mini-repro of homebrew test failure
Summary:
This diff adds a test that repros the weird behavior we see in https://github.com/Homebrew/homebrew-core/pull/160348. The issue is the most likely to trigger on mac, where the default tmp dir `/tmp` is actually a symlink to `/private/tmp`, but the underlying issue can be triggered on all OSes. We just need to pass in a path that has two parts `${path that's symlinked to another realpath}/${non-existing directory yet}`. This diff adds a test to show the broken behavior, which will be fixed in the next diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53199030

fbshipit-source-id: dd31d3c70ab4e7155a3639f4c0471176525e2b09