[flow] Fix weird behavior of inconsistent behavior when tmp-dir of flow is in a symlink directory
Summary:
What's happening with the test in the previous diff? Well, when we first constructed the options object, we have computed a builtin flowlib target unpack directory that maybe a relative path or not. It is a relative path is the temporary directory doesn't exist yet, and it's an absolute path if the temporary directory already exist. Unless users are modifying flow temp dirs, the likely scenarios of two cases above are first run and second run.

However, the rest of our system generally assumes one version of file path, which is the absolute file path, so we have this dangerous inconsistency for a long time, but we never run into any problems since libdefs are treated in a special way. However, now that libdefs are treated more like normal files, this inconsistency starts to cause problems. We turned on `libdef_in_checking` in 0.226.0 and that's when our homebrew tests first start to fail.

We run into an inconsistent state where the parsing stage will store the file key of libdef files in absolute path, but when we are reading the asts of libdefs, we are using potentially relative path file key to query. We don't find anything, which in normal cases it's a result of libdef files that don't parse, so we exit and print all errors we have. Since there is no actual errors, we exit with 0 errors. Boom!

This diff fixes it by normalizing the libdef path to be absolute when we are calling `Files.init`, which ensures that all the libdefs files we present to the rest of the system will always be in absolute path form. In the long term, to avoid this kind of bugs, we should make use of OCaml's type system to clearly distinguish paths in different normalization state, but for now, I just want to fix the bug ahead of a release.

Changelog: [fix] We fixed a bug that causes the first run of `flow check` to always return 0 errors, when some parts of temporary directory is a symlink.

Reviewed By: panagosg7

Differential Revision: D53204521

fbshipit-source-id: 70cc244789fa1412b5c30e446b5ff3cb490f4270