[flow] Fuse module_info and cjs_exports_state
Summary:
This diff fuses together module_info and cjs_exports state. It does come with some behavior change, but in practice doesn't affect most of the existing code.

In particular, we will only change the cjs_exports state if the module is not decided to be ESM kind. Code affected by this will already have indetermined module error, which is very uncommon.

Changelog: [internal] multiplatform not announced externally

Reviewed By: panagosg7

Differential Revision: D53031648

fbshipit-source-id: b29aac69b5cf9ebd2c1844b102e5e7fff95a9b52