[flow][autocomplete][refactor] combine fields of ServerProt.Response.Completion in new structure
Summary:
The 'documentation' and 'tags' field are computed and are typically considered together, so here they are combined in a pair in a new structure `completion_item` in autocompleteService_js.ml. This reduces some noise from the next diff.

The following diff will make this field lazy. This is the reason we need this structure to be separate from `ServerProt.Response.Completion` as we can't store lazy values in that one.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D52670619

fbshipit-source-id: c9ae9aa3f08799d33fc39c26159c4b179245fee3