[flow][autocomplete] make documentation and tags lazy
Summary:
Autocomplete computes documentation and tags by calling into `GetDef_js.get_def` for each candidate name. The call to `get_def` can be expensive since it potentially computes scope information for the entire AST. At the same time it is not guaranteed that we'll be using the information associated with a certain name, since the list of names is filtered out, based on various filters later on.

Instead of always computing doc and tag information, this diff makes the `documentation_and_tags` field lazy. This is only forced in commandHandler.ml, when the result is finalized and is sent to over to the client.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision:
D52670618

------------------------------------------------------------------------
(from 9b73c126d181c6bfa3aec018f0d0ced5c92c3141)

fbshipit-source-id: 7131fa4bdb634cd8de1396143f203580df3768c6