[flow-upgrade][easy] Just use `jest` as the command for `yarn test` in flow upgrade
Summary:
If `yarn test` is set to run `jest --watch`, you have to do that, but if it is set to just `jest`, you can choose where to do that or do `yarn test --watch` which will run `jest --watch`.
I typically don't want to run in `--watch` mode.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D53239997

fbshipit-source-id: 733a4f4d7081a271ea0ac74b2b2e490b9f76b2e4