[flow] Initial rough declare namespace support (excluding type-sig)
Summary:
In this diff, we added type checking support for namespaces. The behavior is not completely at where I want it to be in at least the following ways:

1. Uses import_export mechanism to generate types, but TS namespaces don't require you to export anything: [example](https://www.typescriptlang.org/play?#code/CYUwxgNghgTiAEA7KBbEBnADlMCBGA9gC4AWA+kQJ6YhlSLBkBuUEArggN4BQ88IAD0wEYReGAKJ0YvLACMALiRsUeEDADcvcZOnxZMAExLCBCCHpa+VGvABCUAF7wAvPGkwAlogDmWgL7c3ISkFNS09Iws7CAAdAZy7lBEnugAZp4YSCBM6hrwAPQF-DAwIkqIKmow8AB+AHz8KJhUwcTkNhEMzKwc8bCGSSnpmej67WE0dN3RfQ6OWkA)
2. We still rely on using the parser to reject most of the statements that don't belong in declare namespace.
3. No support for directly exporting a declare namespace

However, as shown by the tests, the feature is mostly there and pretty usable now. Notably, I also included a type checking time check that rejects `declare module.exports` inside declare namespaces, since the equivalent in TypeScript `exports =` is disallowed.

I intend to start with this and continuously iterate on the implementation.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D53107519

fbshipit-source-id: 6796bbe56168e4577f744d4d7a2123cca9256f35