[flow] Ban `declare export type ...` in toplevel context in type checking stage
Summary:
We ban `declare export type` in toplevel statements not for any limitations, but mostly because `declare` is redundant here. In previous diffs, we have since allowed `export type` in declare module and namespaces, it's the time to remove this syntactic limitation as well.

This diff moves the validation of this redundant syntax to type checking stage. The goal is to fully deprecate the syntax eventually.

Changelog: [parser][errors] `declare export type` and `declare export interface` statements in toplevel are no longer parser errors. Instead, they will now be errored during type checking.

Reviewed By: mvitousek

Differential Revision: D53206262

fbshipit-source-id: ac786b2603542d77b948886e9a817ffb96547902