[flow][type-sig] Include enums and components as candidate for auto-export in declare module
Summary:
I plan to reuse the same code for declare namespace's auto-export behavior, then I find that we are missing some cases. Enums and components are not auto-exported when we cannot determine the kind of declare module.

Changelog: [fix] In `declare module`, all the values will be auto-exported if we cannot decide whether the module is in CJS or ESM. However, we are missing the cases for enums. We fixed the issue.

Reviewed By: mvitousek

Differential Revision: D53237693

fbshipit-source-id: ef6a6772994b660c7bf6d1b8960894f9fa9ac117