[flow][refactor] Further simplify namespace type generation
Summary:
Building on top of D53243915. we can further simplify the creation of a namespace type. Instead of creating a module type and then convert it into namespace form, we directly create the namespace form. In the future, when we no longer depend on the TypeT hack, the helper function created here should directly generate a namespace type.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D53278347

fbshipit-source-id: 5693f2b7f11e191e4e7e1091d99244bdb8850dbb