[flow][type-sig] Include every binding as candidate for auto-export in declare module
Summary:
Another followup after D53237693. It turns out that we are also missing declare let/const support for auto-exported names in declare module. To prevent future mistakes like this, I make the pattern matching exhausive. I also included some unsupported statements like const ref (which could be supported in the future), and added test to assert the current behavior.

Changelog: [fix] In declare module, all the values will be auto-exported if we cannot decide whether the module is in CJS or ESM. However, we are missing the cases for `declare const` and `declare let`. We fixed the issue.

Reviewed By: mvitousek

Differential Revision: D53277331

fbshipit-source-id: 22bb7dc5ed94a16e74c12cd6c99e2b8ff64a42f0