[flow] Handle the `use` hook correctly
Summary: The `use` hook has some special rules that make it different from other hooks. It can be called conditionally, and at least according to the existing eslint regression tests, it can be called within a callback. (I don't understand why this is the case, but see `App5` in the examples, which comes direct from the expected passing test cases of the eslint.)

Reviewed By: SamChou19815

Differential Revision: D53286989

fbshipit-source-id: 2d36f6c2c09cd4b4a67524ad4b451ecccd614be6