[flow] Simplify the type reference transform from React.AbstractComponent to React.Element
Summary:
This diff aligns the behavior with annotation inference, where we don't pass in the props type argument, and let the type app does its job of computing the default on its own.

It shouldn't affect the behavior of user code, since the underlying computation is the same.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53333988

fbshipit-source-id: f75289608b6038861476e6a5dd7676f863391c76