[flow] Use a specialized instantiation for PolyT(AbstractComponentT) -> TypeT transform
Summary:
We call `instantiate_poly_with_targs` with quite a lot of checks that will be trivially passing, since we provided the targs as a list of any that has the same size as tparams. This diff replaces that with a simple substitution, to remove dependency on flow_js and make future refactors easier.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53337094

fbshipit-source-id: 1c2eea8a85e48a31164d65db455b395af4af7197