[flow] Add component_syntax flag to flowconfig
Summary:
For the eventual OSS announcement, this new component_syntax flag turns on:
1. component and hook syntax
2. deep read only props
3. deep read only hook returns
4. rules of hooks
5. ref access during render validation

If anything else should be enabled, like hook-like functions, or if anything should be removed please let me know.

Note that this diff is purely additive, both to give us the ability to enable things selectively internally until the implementations settle and to not break existing usages of this flag. When everything settles let's remove the other flags.

Changelog: [internal]

Reviewed By: alexmckenley

Differential Revision: D53355340

fbshipit-source-id: 1723bd6776c7016f8095879957f7e5ba52b46d7d