[flow][casting-syntax] Make `both` the default value for the `casting_syntax` option, if not supplied
Summary:
Make `both` (allowing both `as` expressions and the old syntax) the default value for the `casting_syntax` option, if not supplied.
This will be in the release after the open-source announcement.

The `as_const` option now works without a `casting_syntax` option specified, as the default for the latter `both`.

Reviewed By: SamChou19815

Differential Revision: D53458521

fbshipit-source-id: f5e384794113b11e213a156682342c3610bb8071