[flow][test] Add test demonstrating inconsistent behavior of using TS type names
Summary:
Currently, we will ban the usage of types like `Readonly` with `ts-syntax` errors, even if user adds a shim in the global libdef like `
type Readonly<T> = $ReadOnly<T>;`.

At the same time, it turns out that we are also inconsistent with the treatment of such types. In local type checking, the type will be turned into any, while in type_sig, we will happily read from the global.

This diff adds the test to demonstrate the issue, which will be addressed later in the stack.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53371950

fbshipit-source-id: 517baea9d941ec49f681de443d63f8f9ac83e547