[flow] Introduce `experimental.ts_syntax` config to allow some TS syntax
Summary:
Changelog: [feature]

We introduced `experimental.ts_syntax` flag that you can set to `true`. When this flag is on, Flow will no longer complain about certain TypeScript syntax, and it would automatically turn them into the closest Flow equivalent types.

Initially, we support the following: `Readonly`, `ReadonlyArray`, `ReadonlyMap`, `ReadonlySet`, `NonNullable`, and using `extends` to specify type parameter bounds. In the future, more kinds of types that we currently parse but error on might be added. (e.g. keyof)

Please note that the support here is best effort. We make no guarantee that these types have the same semantics as those in TypeScript 100% of the time, but for common cases, it might be good enough for you.

Reviewed By: gkz

Differential Revision: D53371951

fbshipit-source-id: 910454c8148b955cf6349c02edde7dec6e38b1df