[flow] only raise [ambiguous-object-type] when 'exact_by_default=false'
Summary:
The [ambiguous-object-type] error is not enabled by default but internally it is always raised on object types that have neither the exactness bars (`{||}`) nor the inexactness ellipses (`{...}`). Even when the `ambiguous-object-type=error` lint is set, this error should not be raised when exactness is enabled by default, since the `{}` form is the one recommended for an exact object (since we released exact by default by default). The wording of the error: "(use `{|` and `|}` instead of `{` and `}`)" is also not compatible with exact by default.

This diff makes this error only be raised when the `exact_by_default` option is `false`.

Changelog: [errors] the [ambiguous-object-type] error will only be raised when `exact_by_default=false`.

Reviewed By: SamChou19815

Differential Revision: D53500030

fbshipit-source-id: cf623107028288f4d99230e092a414753e980e24