[flow] Recognize `.current` on a DRO object as a read of a React ref
Summary:
Currently, the validateRefAccessInRender rule only fires on reads from the builtin `React$RefObject` type. This is easy to bypass simply by flowing a ref to a value with the type `{ current: T }`. A component could, for example, take a ref as a prop with such a type.

This diff (partially) addresses this issue by noting that props, hook returns, and hook arguments all are treated as deep-read-only, which is manifested as a flag in the ObjT. If that flag is set, we can then see if the object only has a .current property, and if so, treat it as a ref. This introduces a possibility for false positives, for example if a component takes an object like { current: T } and uses it in a totally non-reflike way, but it seems unlikely.

In addition, this changes the React libdef to use `React$RefObject` as the return type of `React.createRef`, which should help catch a few more cases.

Reviewed By: jbrown215

Differential Revision: D53437727

fbshipit-source-id: 15861dd2a3862ecd5ead9a6be245b880f1f971e9