[flow] ensure_parsed before merging for codemods to enable the usage of saved state
Summary:
Right now, if we try to use saved state for glean indexer or any other codemod that uses the typed runner, it will crash with some "expect <some file> to be parsed" exception. It has less to do with saved state, but with the implicit assumption in the current code that all the files are care about is already parsed, which is guaranteed if we do an init without saved state.

This diff fixes it by calling `ensure_parsed`, similar to what we do for normal flow server. Now we can run codemods and glean indexers with saved state.

Changelog: [fix] We now allow the use of saved state for [glean](https://glean.software/) indexer, if saved state is enabled in flowoconfig or CLI arguments.

Reviewed By: panagosg7

Differential Revision: D53534974

fbshipit-source-id: 444f2bf196c994239647862443d3d918f330c60f