[flow] do not create PredicateT (LatentP) on non-predicate callees
Summary:
Before, on every call that happened to appear in a conditional context, e.g.
```
if (foo(x)) { ... }
```
we would introduce a `LatentR` refinement on `x` in the name resolver. Later on when populating the Type_env, the `LatentR` would get translated to a `LatentP` type predicate on `x`. This would happen regardless of whether `foo` was a predicate function or not.

This indirection is unnecessary and problematic since it forces the incoming type of `x` to go through the `PredicateT` mechanism. Specifically when the type of `x` was a large union, it would cause the union to be broken up to its parts, causing slowdowns.

This diff tries to address this by disabling the `LatentR` -> `LatentP` conversion when the callee is not a predicate function type. For that case, which is by far the most common case, we introduce a new predicate `NoP` which acts as a no-op. Crucially, the constraints:
```
_ ~> PredicateT (NoP, u)
_ ~> PredicateT (NotP NoP, u)
```
are handled very early in the pattern match so that the LHS is propagated to `u`, before any unnecessary processing happens.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D50204046

fbshipit-source-id: b08c714aac5a05c40a33fb5de1ef7ac256064ace