[flow] make PredBased predicate maps lazy
Summary:
The purpose of this change is to prevent an error in the next diff due to computing the type of the following function
```
function foo(x: mixed): boolean %checks {
  return foo(x);
}
```
Specifically, to produce the full type of `foo` we need to determine if the call `foo(x)` is a predicate call. This will help us decide whether we need to introduce a `LatentP` or not. But determining this means that we need the type of `foo`. This introduces a read on the type of `foo` before it has been computed, which returns `empty`.

So, instead, we make the predicate part of `foo` lazy, so that by the time we need to access the predicate the type of `foo` has been installed in the type env.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D53600508

fbshipit-source-id: 065fcadb628fc9f71d64ad2e032cf4f91d21a47f