[flow] Use simple repos_reason for `Type_annotation.identifier`
Summary:
It appears that we don't need a reposition here, since the equivalent place in type sig also doesn't do it. All we need is replacing the toplevel reason, possibly on OpenT, so that some other non-termination defenses based on unique reasons can work.

With this, we no longer do any repositioning in type_annotation.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53554103

fbshipit-source-id: 0c056f4dbd999bc48167cb27f244fa1eddf23971