[flow][autoimports] Only include autoimport results during autocomplete if user typed certain mininum number of characters
Summary:
Say the user only typed 2 characters for some identifier to autocomplete. Are we going to get any useful autoimport results? Probably not. However, for a large codebase with millions of exported symbols, we might waste a lot of time searching through the index.

In this diff, I introduced an option to let you configure the minimum number of characters you need to type in order to include any auto import results. It's defaulted to 0 for now to preserve the existing behavior, but we will experiment internally to see how much we can push it without degrading autocomplete acceptance rate.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53547698

fbshipit-source-id: 4c2713f858cb8eac74b441ac8676bf53c4f6848b