[flow] Basic support for enforcing union optimization
Summary:
Determining if Flow has optimized a union type is currently not practical. This diff tries to change this by adding a new type `$Flow$EnforceOptimized` that takes the target type as argument. Flow will raise an error if the argument type is not union-optimizable. This utility type has otherwise no other effect and is equivalent to the type argument it is passed.

It's implementation tries to match, as much as possible, the code path we follow when optimizing unions. A union is optimized when it appears as an upper bound to a constraint, with the hope that it might prevent expensive speculation jobs. To mimic this, whenever we have a `$Flow$EnforceOptimized<T>` type, we create a constraint between a new internal type and `T`.

In this diff, we raise a generic error if:
* the input type is not a union
* the optimization fails to produce a result, or
* the result is `Unoptimized`.

In the next diff we will add more detail on the third case, explaining exactly what went wrong.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D53634989

fbshipit-source-id: d792df8fc37f59d919c8904c53d7dcd092bfc194