[flow][refactor] Eliminate `Builtins.get_builtin`
Summary:
I plan to separately store the names, types, and modules in globals and provide 3 functions to query them, so that we can avoid the errors of reading a type as value by construction.

Before that, I first want to reduce the API surface by eliminating `Builtins.get_builtin` that doesn't save a lot of effort, but will force me to produce 3*2 functions instead of 3.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53634091

fbshipit-source-id: 8cfd9cf58813323349e30f6306ef772be2ba0bb1