[flow] Always decide whether we are in declare_module from environment
Summary:
Before, we don't have a special scope for declare module, so we used another special flag and record it and pass it into a few places. Now that we have a special scope, we can just rely on the current infra of env_resolution and statement.ml setting the right scope, so that we can always query from the context.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53634094

fbshipit-source-id: 358cf8908d21fd02816cfad4b5cd73f51f77eb7f