[flow] Fix misc errors with hook compatibility mode
Summary:
This diff addresses the remaining known sources of bogus react-rule-hook errors under compatibility mode.

First off, there were a few places where declarations with hooklike names weren't getting the hooklike flag correctly, notably destructured props/array elts and component props. To support these cases, we restructure and simplify the name def handling of hooklike-ness, by creating a new binding kind that makes its argument hooklike. This is similar to a selector, but it doesn't need to store things in the environment the way selectors do. With this binding kind, we can consistently apply hooklikeness based on naming on all values introduced with bindings (in the name def sense, so variable declarators and parameters, but not function declarations, which still must be handled separately).

The second issue was that applying HooklikeT to an intersection wasn't making the elements of the intersection hooklike. It'd be fairly complicated to correctly apply hooklikeness to intersection elements and then rebuild the intersection in such a way that it'd be usable by function overload resolution, so instead we just punt on intersections when the compatibility mode is enabled. As an additional twist to this, the react rule analysis needs to know that a callee is an intersection, but the type recorded in the typed AST is the resolved overload, which is not an intersection. To handle this, we look into the environment rather than the AST to get callee types. This will still fail on intersection callees that are expressions rather than identifiers, but those don't appear to be used much if ever in this context.

Finally, there was an issue that a looked-up object prop would never get both DRO'd and hooklike'd because those flags are applied in the same match statement. We now split up that match so both can be applied if appropriate.

Reviewed By: SamChou19815

Differential Revision: D53594462

fbshipit-source-id: 9e9eec2bdd29cdb4e494be6a5948aa5d019fbceb