[flow][refactor] Separately store names and modules in builtins
Summary:
Now that we have shown that we will never lookup a global as either name or module in one place, we can start to store them separately. This alone isn't that important, but it will pave ways for future refactors that store values and types separately, with the eventual goal of killing the TypeT wrapper.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D53634084

fbshipit-source-id: 01a883239dbe30a39837b724cbe780ce7f2e9a59