Remove blocking reads/writes from MultiWorkerLwt (redux)
Summary:
This diff re-introduces non-blocking MultiWorker reads and writes. The original
change was reverted because we did not handle cancelation during send correctly.

This diff adds handling for cancelation. First, we protect the send to ensure
the entire job is received by the worker. Currently, a worker treats an
incomplete send as a fatal error. However, if the worker receives a full job, it
will start processing it and eventually notice the cancelation.

Original commit changeset: 5baff688b40b
Original Phabricator Diff: D52339137

Reviewed By: SamChou19815

Differential Revision: D52345242

fbshipit-source-id: 5de2fb04bdbbd0d060a8539024e3d5243ab4c61d