[flow] Separate values and types in type_sig
Summary:
In this diff, we partition the values and types into two different maps in type sig's scope analysis, which allows us to output two different sets of globals: one contains only values (including ones that can be used as type references according to our rules), and another set of type-only stuff (type aliases, interfaces, ...).

In the type sig analysis, we ensure that the map of values and the map of types are disjoint. When we do lookup, we are looking for value xor type, instead of either is ok. When we look for values, we won't look at the type map at all. When we look for types, we look for both.

This changes prepares us to better reject code like `const x = React$MixedElement` in the future (currently no error, until you try to use it and get a confusing subtyping error), and also prepares us to move away from the `TypeT` wrapper.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53688498

fbshipit-source-id: 99515a0615826453c57306de5a4f469bc009a5e4