[flow] Directly mod reason in flow_js_utils's global lookup function
Summary:
Currently, we only do this in flow_js, which used to be a full reposition, but it turns out that modding the reason is sufficient. Since we don't actually need reposition, we should just consistently do it in flow_js_utils, which will enable us to delete a lot of slight variants of global lookup functions that's often unclear which one should we use.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53733946

fbshipit-source-id: 2e6450021e62528df58b6e0b925e06271593eb8c