[flow][refactor] Directly error on assignment to unknown global
Summary:
Over here, we do a look up not for the sake of getting the type, but to trigger a potential unknown global error. We should just error directly. I also changed it from a general name lookup to value lookup, since assignment to a type global is also not good...

Changelog: [errors] Previously, we allow you to assign to any globals, regardless of whether it's a type or value. Therefore, both are allowed in `Array = 3; $ReadOnlyArray = 2;` Now the latter is banned.

Reviewed By: panagosg7

Differential Revision: D53734245

fbshipit-source-id: 86e9edab83ab776facaad2f8e89e23858f4f896f