[flow] Handle global read of values and types separately
Summary:
In this diff, we eliminate the last place in type checking code where we either read a global value or global type. To accompolish this, we need to keep track of the kind of globals we are expecting to read.

In type_sig, this is achieved by propagating the `type_only` in parsed type sig ref to packed type sig ref. In local checking code, we use the `LookupMode` information.

Changelog: [errors] Flow now consistently bans referring a type-only global as value

Reviewed By: jbrown215

Differential Revision: D53740975

fbshipit-source-id: 8a13e26d809205512c30b7247702c86959125fd7