[flow][test][EZ] Add test exposing inconsistency between local check and type sig wrt typeof other params
Summary:
I am looking at how type sig manages to handle destrucuturing in parameters, given that we allow things like `function foo({bar}: {bar: string}, baz: typeof bar) {}`. It turns out that it doesn't, and it doesn't even bind function parameters..., so in the example above, `typeof bar` will resolve to the `bar` in the upper scope...

The inconsistency was introduced in D46165727, so with hindsight we should probably hold in until we have similar behavior in type_sig.

This diff doesn't fix it, since fixing it is quite involved. It just adds a test to document this bad behavior.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53839920

fbshipit-source-id: 34002c5ba873c61af0866d20e233d74cfa48dea5