[flow][refactor] Group post inference checks into post-component and post-merge
Summary:
Now that we have LTI, we already have the guarantee that a type created in strongly connected component A is not going to get any new bounds after the strongly connected component is resolved. Therefore, we can make most of the post-inference check post-component.

But why we want to do this, given that the current setup is fine? A minor reason is to avoid throwing too many stuff into context that cannot be garbage collected. It might explain why we have memory regression under long-lived worker mode. However, another reason is that I want to delay all the eager evaluation of `EvalT` while we are generating types for a strongly-connected component (for the purpose of LTI 2.0). Sadly, when we do the evaluation of `EvalT` matters (both for error message description and hint decomposition), so it will be good to delay it, but as close to the current time as possible. Post-component is a good phase for that.

This diff does none of that. It only separate what can be done post-component and post-merge. It's done separately to ensure that we are not depending on the order of these operations.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53969726

fbshipit-source-id: 17730cd959de1792f9114ceb9df6ef32e6cd2c46