[flow] Defer more error related to bad annotation in speculation
Summary:
In this diff, I added 3 new errors into `defered_in_speculation`. These errors are results of bad annotation, so the error should be surfaced unconditionally, and shouldn't affect speculation.

The current model of using error kind to decide whether it should throw in speculation is very flawed and needs some overhaul.  (consider the type `namespace.does_not_exist`, we would get prop-missing error, but it's not safe to make prop-missing errors in general deferred in speculation) This diff doesn't attempt to fix it completely. Instead, it just tries to move thing into the right direction to reduce the error diff in the future.

Changelog: [errors] Some bad annotations that cause `value-as-type` or `recursive-definition` errors will no longer affect which branch to take in overload resolution.

Reviewed By: panagosg7

Differential Revision: D54031628

fbshipit-source-id: 5e8e668bf208b2aaaa4d73ba1cab31e70f15fd45