[flow] Introduce `NamespaceT` and `GetTypeFromNamespaceT`
Summary:
This diff creates the boilterplate to setup a new type `NamespaceT`, to represent a bag of both values and types. It's intended to be used to accurately represent the type of namespace import and `declare namespace`, rather than using the current object type hack, where both types and values exist in `props_tmap` and are distinguished by `TypeT` only.

Paired with this, I also introduced `GetTypeFromNamespaceT`, which aims to replace the GetPropT for getting a type out of a namespace.

None of the functionalities of the new type and use_t is implemented in this diff.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D53834702

fbshipit-source-id: 0ec12378ea8c9472499cf1dcf09556c56ecbbd8e