[flow] Implement behavior of `NamespaceT` and `GetTypeFromNamespaceT`
Summary:
In this diff, the behavior of the type and the use_t is implemented. To get a type-only stuff from `NamespaceT`, the only way is through `GetTypeFromNamespaceT`, since all other operations on `NamespaceT` will cause the type to be concretized to the value-only part. I have replaced `ConsGen.get_prop` for qualified types with `ConsGen.quality_type` with this new use_t. I also replaced all the places where we are generating a namespace-like objects with the actual namespace type, so that we start to use this new infra.

I also needed to fix something in normalizers and members.ml that are powering various IDE services, which still assume these special objects. For now, I changed the places that also special case modules to special case namespaces as well, but for the most general case in normalizer, I just convert the namespace type back into an Ty.Obj that contains both values and types. Cleanup of that should come another day, when we start to use the new infra to completely kill `TypeT`.

I would hope I could just stop here, but it turns out that we also need to ensure that the exports are separated by values and types. Currently, we put them into the single map, which might overwrites each other sometimes. (See the example in cjs_esm_interop added in D53861043). We must do the namespace type change and the separation of value/type export in the same diff, so that we have a consistent view of separated value/type everywhere, rather than doing them separately and temporarily introduce more implementation-defined random overwriting behavior.

All these works combined, we finally ensured that we disallow using type as value **by construction**.

Changelog: [errors] We now put value exports and type exports into two different namespaces. It means that we will now consistent error on places where you try to import, re-export, or use a imported type as a value. This change also makes certain patterns that happen to work in the past no longer works, e.g.

```
// a.js
export type Foo = string;
// b.js
const A = require('./a');
module.exports = {...A};
// c.js
// previously allowed, no longer allowed
import type {Foo} from './b';
```

Reviewed By: panagosg7

Differential Revision: D53834701

fbshipit-source-id: ec23021254123c4cb4d2424481843b33ccbf4ab0