[flow] Kill now defunct guard of TypeT in object props
Summary:
Now that a namespace like object actually has a namespace type. It is immediately reduced to the value part in normalizer, unless its in toplevel, where we have special treatment of exported types. Thus, it's not guaranteed by construction that we will never see `TypeT` in object types in normalizer, so we can delete this guard.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54095386

fbshipit-source-id: 3e136aa0e30da2bfae61e32c62fd656a7a692902